+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- feat(git): add support for reading remote URLs and enhance git connection management. (`92b43d4`)
+
 ## 2.3.0 - 2026-07-13
 
 - feat(dependencies, themes): update harborclient packages and enhance theme management. (`d5562a1`)